% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bBluff.R
\name{bBluff}
\alias{bBluff}
\title{Create a list of buffers from a multi-line shapefile.}
\usage{
bBluff(coast1, dir.name)
}
\arguments{
\item{coast1}{\itemize{
\item Either a SpatialLinesDataFrame object or a full path and file
name including extension .shp for a multiline shapefile.
The shapefile has to be projected, not in lat/long geographical coordinates.
The shapefile should have in attribute table at least an unique ID column
named ID and a column named buff_dist that has the distance desired for buffer
in adequate measurement units from the file projection
(usually meters for a UTM projection).
}}

\item{dir.name}{\itemize{
\item The full path to the directory where a Buffers subdirectory
will be created and the individual buffers will be saved. If the dir.name given
does not exist it will be created. If dir.name is NA, the buffers will not be saved
as polygon shapefiles on the disk.
}}
}
\value{
A list of projected SpatialPolygonsDataFrame.
}
\description{
Create a list of buffers from a multi-line shapefile.
}
\examples{
coast1 <- system.file("extdata", "shore_line.shp", package = "iBluff")
buff1 <- bBluff(coast1, NA)
}
