\name{BDEsize-internal}
\alias{BDEsize-internal}
\alias{fsize}
\alias{sizelist}
\alias{sizelist.split}
\title{Internal BDEsize Functions}
\description{
  The \code{fsize}, \code{sizelist} and \code{sizelist.split} are internal BDEsize functions.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
