% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bProf.R
\name{bProf}
\alias{bProf}
\title{Extract elevation profiles}
\usage{
bProf(lns1, rast1, buff1, dir.name)
}
\arguments{
\item{lns1}{\itemize{
\item Is a list of class "Lines" and is the result (the 2nd list term) of the
bTrans function for each considered buffer.
}}

\item{rast1}{\itemize{
\item Either full path including file name and extension (e.g. .tif for geotiff files)
of one DEM raster that was clipped with a buffer or a RasterLayer object.
}}

\item{buff1}{\itemize{
\item Either full path including file name and extension .shp of one
polygon buffer shapefile, or a SpatialPolygonsDataFrame object that contains
one polygon shapefile. This is usually the result from bBluff function, one
object from the resulting list.
}}

\item{dir.name}{\itemize{
\item The full path to the directory where a bluff profile subdirectory
called Profiles will be created.The Profiles subdirectory will have separate
subdirectories for each buffer. Individual comma delimited tables (*.csv) will be
saved for each extracted profile. If the dir.name given does not exist it will be created.
If dir.name is NA, the *.csv tables will not be saved on the disk.
}}
}
\value{
A list of lists of data.frames, each table is a bluff/cliff profile.
Each table in the list has 6 columns in this order with names:
xi, yi, zi, id, id1, and ID.prof where:
\itemize{
\item xi, yi, zi - coordinates, where zi represents elevation (in meters);
\item id - is a sequential id of rows in the table;
\item id1 = is the cumulative distance between points on profile;
\item ID.prof = is the index of the profile, if the list of profile tables has more than
one table then this index will identify uniquely the profile.
\item The profiles are always reoriented to have beach and water to the left and
bluff face and land to the right.
}
}
\description{
This function extracts elevation profiles along the transects within the DEM
buffers. The function extracts one single profile in approximately 2 seconds in average
if the profile is also saved on disk.
}
\examples{
# Do not run
# coast1 <- system.file("extdata", "shore_line.shp", package = "iBluff")
# buff1 <- bBluff(coast1, NA)
# shore1 <- bCoast(coast1, NA)
# dem1 <- system.file("extdata", "dem_demo.tif", package = "iBluff")
# rast1 <- clip.rast(dem1, buff1,NA)
#
# n <- length(shore1)
# comb.n <-  combn(1:n,1)
# marg1 <- apply(comb.n, 2, function(x)
#   make.margins(buff1[[x]], shore1[[x]], rast1[[x]], NA))
#
# n <- length(marg1)
# comb.n <-  combn(1:n,1)
# trans1.all <- apply(comb.n, 2, function(x)
#    bTrans(marg1[[x]], dist1=NA, rast1[[x]], NA))
#
# n2 <- length(trans1.all)
# comb.tr <- combn(1:n2, 1)
# lns1 <- apply(comb.tr, 2, function(x) trans1.all[[x]][[2]])
#
# n3 <- length(lns1)
# comb.n3 <- combn(1:n3,1)
# profs <- apply(comb.n3, 2, function(x)
#    bProf(lns1[[x]], rast1[[x]], buff1[[x]], NA))
}
