% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_rast.R
\name{clip.rast}
\alias{clip.rast}
\title{Clip / extract the DEM in the existing buffers}
\usage{
clip.rast(dem, buff, dir.name)
}
\arguments{
\item{dem}{\itemize{
\item The full path and file name including extension (e.g. .tif) of the
raster DEM that will be clipped.
}}

\item{buff}{\itemize{
\item Either a vector (class character) of full path and file names
including extension .shp of the polygon shapefiles, or a list of buffers as
SpatialPolygonsDataFrame objects.
}}

\item{dir.name}{\itemize{
\item The full path to the directory where a DEMs subdirectory
will be created and the individual rasters will be saved as geotiff.
If the dir.name given does not exist it will be created.
If dir.name is NA, the rasters will not be saved on the disk.
}}
}
\value{
A list of rasters.
}
\description{
Clip / extract the DEM in the existing buffers
}
\examples{
coast1 <- system.file("extdata", "shore_line.shp", package = "iBluff")
buff1 <- bBluff(coast1, NA)
dem1 <- system.file("extdata", "dem_demo.tif", package = "iBluff")
dem3 <- clip.rast(dem1, buff1,NA)

}
