% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cols_cntr.R
\name{cols.cntr}
\alias{cols.cntr}
\title{Columns contribution chart}
\usage{
cols.cntr(data, x, cti)
}
\arguments{
\item{data:}{name of the dataset.}

\item{x:}{dimension for which the column categories contribution is returned.}

\item{cti:}{logical value which specifies if the contribution to the total inertia must be displayed as well. No missing value is allowed. Use T for 'true', F for 'false'.}
}
\description{
This function allows you to calculate the contribution of the column categories to the selected dimensions.
The function returns a dot plot in which a reference line indicates the threshold for a significant (i.e., important for the sake of dimension interpretation) contribution.
By selecting one of the function's parameters as true (i.e., cti=T), the plot will also show the contributions to the total inertia, which is represented by hollow circles. This option allows to compare the contribution of each category to the inertia of the dimension in the overall framework of its contribution to the overall inertia of the input contingency table.
}
\examples{
data(greenacre_data)
cols.cntr(greenacre_data, 2, T) # Returns the contribution of the row categories to first 2 CA dimensions, and also displays the contribution to the total inertia.
}
\keyword{columns}
\keyword{contribution}

