% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cols_corr_scatter.R
\name{cols.corr.scatter}
\alias{cols.corr.scatter}
\title{Scatterplot for column categories correlation with dimensions}
\usage{
cols.corr.scatter(data, x, y)
}
\arguments{
\item{x,y:}{dimensions for which the correlations are reported.}

\item{data:}{Name of the dataset.}
}
\description{
This function allows to plot a scatterplot of the correlation of column categories with two selected dimensions. A diagonal line (in BLACK) is just a visual aid to eyeball whether a category is actually more correlated (in relative terms) to either of the two dimensions.
}
\examples{
data(greenacre_data) #load the sample dataset
cols.corr.scatter(greenacre_data,1,2) #plot the scatterplot for dimensions 1&2
}
\keyword{categories}
\keyword{column}
\keyword{correlation}
\keyword{scatterplot}

