% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cols_qlt.R
\name{cols.qlt}
\alias{cols.qlt}
\title{Chart of columns quality of the display}
\usage{
cols.qlt(data, x, y)
}
\arguments{
\item{x,}{y: dimensions for which the correlations are returned.}

\item{data:}{name of the dataset.}
}
\description{
This function allows you to calculate the quality of the display of the column categories on pairs of selected dimensions.
}
\examples{
data(greenacre_data)
cols.qlt(greenacre_data, 1,2) #Returns the quality of the display of the column categories on the 1&2 dimensions.
}
\keyword{cols}
\keyword{quality}

