% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toptoe.R
\name{get.toptoe}
\alias{get.toptoe}
\title{Get top and toe coordinates on an elevation bluff / cliff profile}
\usage{
get.toptoe(profs, top.par = TRUE, veg = FALSE)
}
\arguments{
\item{profs}{\itemize{
\item A list of data.frames, each table is a bluff/cliff profile.
Each table in the list has 6 columns in this order with names:
xi, yi, zi, id, id1, and ID.prof where:
\item xi, yi, zi - coordinates, where zi represents elevation (in meters);
\item id - is a sequential id of rows in the table;
\item id1 = is the cumulative distance between points on profile;
\item ID.prof = is the index of the profile, if the list of profile tables has more than
one table then this index will identify uniquely the profile.
\item The profiles are always reoriented to have beach and water to the left and
bluff face and land to the right.
}}

\item{top.par}{\itemize{
\item A TRUE/FALSE parameter to check that the top elevation
is above the mean range elevation and force the top to be above it. This is useful
in particular situations when the profile has a very exacerbated convex
portion that is not considered the top of the bluff.
}}

\item{veg}{\itemize{
\item A TRUE/FALSE parameter to eliminate 1 point anomalies on the profile if
vegetation was not totally eliminated from the DEM. The default value is FALSE.
}}
}
\value{
A list of four lists of data.frames for each processed buffer.
First list within a buffer list has coordinates and ID for top of bluffs.
Second list within a buffer list has coordinates and ID for toe of bluffs.
Third list within a buffer list has the data used when parameter veg=TRUE.
Forth list within a buffer has the data when parameter veg=FALSE.
}
\description{
The function first determines the position of the toe on the bluff profile in
case of profile concavity, before it gets the top.
}
\examples{
# Do not run
# profs is the list of lists of data.fame of profiles.
# See the example for the bProf function.
#
# n4 <- length(profs)
# comb.n4 <- combn(1:n4,1)
#
# toptoe1 <- apply(comb.n4, 2, function(x)
#   get.toptoe(profs[[x]], top.par=TRUE, veg=TRUE))
#
# n5 <- length(toptoe1)
# comb.n5 <- combn(1:n5, 1)
#
# list of bluff tops coordinates
# top1 <- apply(comb.n5, 2, function(x) toptoe1[[x]][[1]])
#
# list of bluff toes coordinates
# toe1 <- apply(comb.n5, 2, function(x) toptoe1[[x]][[2]])
#
# list of profiles with small errors removed
# prof.noveg <- apply(comb.n5, 2, function(x) toptoe1[[x]][[3]])
#
# list of original profiles
# prof.orig <- apply(comb.n5, 2, function(x) toptoe1[[x]][[4]])
#
}
