% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBeach.R
\name{iBeach}
\alias{iBeach}
\title{Extract the beach profile}
\usage{
iBeach(prof1, toe1)
}
\arguments{
\item{prof1}{\itemize{
\item One profile data.frame within a buffer with the following
columns in this strict order: "xi", "yi", "zi", "id", "id1", "ID.prof"
}}

\item{toe1}{\itemize{
\item Is a list of 1 row data.frame with same columns as the prof1 that
correspond to the toe of the bluff/cliff in all existing profiles within a buffer
DEM (result of get.toptoe function). In prof1 and toe1 the ID.prof (which is the 6th
column always) corresponds to the profile ID for the toe, and the profile itself.
}}
}
\value{
A data.frame with x, y, z coordinates for points on the beach, id as unique
identified, id1 as cumulative distance between points and ID.prof as unique
identifier of the profile it came from.
}
\description{
The function extracts the beach profile if the beach exists from a profile where
the bluff/cliff toe position is known.
}
\examples{
# Do not run
# Use the prof.noveg list of data.frames and the bluff / cliff top/toe list (top1/toe1) for
# obtaining the bluff/cliff face profile. See get.toptoe function example. Or use only
# the prof.noveg list of data.frames is the bluff=FALSE for which top1=NA and toe1=NA.
#
# n7 <- length(prof.noveg)
# comb.n7 <- combn(1:n7,1)
#
# profs1 <- apply(comb.n7, 2, function(x)
#   {n8<- length(prof.noveg[[x]]);
#    comb.n8 <- combn(1:n8,1);
#    apply(comb.n8, 2, function(i)
#      iBeach(prof.noveg[[x]][[i]], toe1[[x]]))})
#
}
