% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBSecond.R
\name{iBSecond}
\alias{iBSecond}
\title{Secondary inflections on bluff/cliff face}
\usage{
iBSecond(profs, top1, toe1, iHi)
}
\arguments{
\item{profs}{\itemize{
\item A list of data.frames, each table is a bluff/cliff profile.
Each table in the list has 6 columns in this order with names:
xi, yi, zi, id, id1, and ID.prof where:
\item xi, yi, zi - coordinates, where zi represents elevation (in meters);
\item id - is a sequential id of rows in the table;
\item id1 = is the cumulative distance between points on profile;
\item ID.prof = is the index of the profile, if the list of profile tables has more than
one table then this index will identify uniquely the profile.
\item The profiles are always reoriented to have beach and water to the left and
bluff face and land to the right.
}}

\item{top1}{\itemize{
\item A list of bluff/cliff top data.frame with one row and same column
names as the data.frame from profs list. This is the results from get.toptoe function.
}}

\item{toe1}{\itemize{
\item A list of bluff/cliff toe data.frame with one row and same column
names as the data.frame from profs list. This is the results from get.toptoe function.
}}

\item{iHi}{\itemize{
\item Is a threshold in meters that will classify top2 and toe2 as
significant if their detrended elevation is equal or greater than threshold.
}}
}
\value{
A list of lists with two elements. The first element is a list of secondary tops data.frame (top2).
The second elemnt is a list of secondary toes data.frame (toe2). The data.frames have the following columns:
xi, yi, zi, id, id1, ID.prof, len, v8, signif1 where:
\itemize{
\item xi, yi, zi - coordinates, where zi represents elevation (in meters);
\item id - is a sequential id of respective secondary top/toe row in the original profile table;
\item id1 = is the cumulative distance between points on profile that corresponds to the secondary top/toe;
\item ID.prof = is an index of the profile, if the list of profile tables has more than
one table then this index will identify uniquely the profile.
\item len represents the detrended value of the elevation for secondary top/toe;
\item v8 is the sequential id of respective secondary top/toe row in the bluff/cliff face profile table;
\item signif1 is 0 for not significant and 1 for significant.
}
}
\description{
The function looks for the major "positive" and "negative" inflections on the face
of the bluff/cliff profile between top and toe.
}
\examples{
# Do not run
# Use the prof.noveg list of data.frames and the bluff / cliff top/toe list (top1/toe1) for
# obtaining the bluff/cliff face profile. See get.toptoe function example.
#
# n10 <- length(prof.noveg)
# comb.n10 <- combn(1:n10,1)
#
# sec.toptoe <- apply(comb.n10, 2, function(x)
#   iBSecond(prof.noveg[[x]], top1[[x]], toe1[[x]], iHi=1))
#
# n11 <- length(sec.toptoe)
# comb.n11 <- combn(1:n11,1)
#
# top2 <- apply(comb.n11, 2, function(x) sec.toptoe[[x]][[1]])
#
# toe2 <- apply(comb.n11, 2, function(x) sec.toptoe[[x]][[2]])
#
}
