% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iRastCC.R
\name{iRastCC}
\alias{iRastCC}
\title{Create a concave/convex raster of the bluff/cliff face or buffer}
\usage{
iRastCC(profs, top1 = NA, toe1 = NA, buffdem, bluff = FALSE, dir.name = NA)
}
\arguments{
\item{profs}{\itemize{
\item A list of data.frames, each table is a bluff/cliff profile.
Each table in the list has 6 columns in this order with names:
xi, yi, zi, id, id1, and ID.prof where:
\item xi, yi, zi - coordinates, where zi represents elevation (in meters);
\item id - is a sequential id of rows in the table;
\item id1 = is the cumulative distance between points on profile;
\item ID.prof = is the index of the profile, if the list of profile tables has more than
one table then this index will identify uniquely the profile.
\item The profiles are always reoriented to have beach and water to the left and
bluff face and land to the right.
}}

\item{top1}{\itemize{
\item A list of bluff/cliff top data.frame with one row and same column
names as the data.frame from profs list. This is the results from get.toptoe function.
\item If the bluff parameter is FALSE (default) then top1=NA (default).
}}

\item{toe1}{\itemize{
\item A list of bluff/cliff toe data.frame with one row and same column
names as the data.frame from profs list. This is the results from get.toptoe function.
\item If the bluff parameter is FALSE (default) then toe1=NA (default).
}}

\item{buffdem}{\itemize{
\item Either full path including file name and extension (e.g. .tif for geotiff files)
of one DEM raster that was clipped with a buffer or a RasterLayer object.
}}

\item{bluff}{\itemize{
\item A TRUE/FALSE parameter that indicates if the concavity/convexity is
calculated on the original bluff/cliff profile (Default: FALSE) or on bluff/cliff
face between top and toe (TRUE).
\item If bluff=TRUE, then top1 and toe1 parameters are lists of top and toe data respectively,
they cannot be NA.
}}

\item{dir.name}{\itemize{
\item The full path to the directory where the subdirectory "ConvexConcave_rast"
will be created to save the raster in geotif format. If the dir.name given does
not exist it will be created. If dir.name is NA, the raster will not be saved on the disk.
}}
}
\value{
A geotiff raster for each processed bluff/cliff buffer.
}
\description{
The function creates along bluff/cliff face profiles (between top and toe) a relative
convex (positive values) / concave (negative values) raster, when the bluff parameter is TRUE,
otherwise the convex/concave raster is created along the original profile length.
}
\examples{
# Do not run
# Use the prof.noveg list of data.frames and the bluff / cliff top/toe list (top1/toe1) for
# obtaining the bluff/cliff face profile. See get.toptoe function example. Or use only
# the prof.noveg list of data.frames is the bluff=FALSE for which top1=NA and toe1=NA.
#
# n10 <- length(prof.noveg)
# comb.n10 <- combn(1:n10,1)
#
# For bluff faces
# cc1<- apply(comb.n10, 2, function(x)
# iRastCC(prof.noveg[[x]], top1[[x]], toe1[[x]], rast1[[x]], bluff=TRUE, NA))
#
# For the entire profile
# cc2<- apply(comb.n10, 2, function(x)
# iRastCC(prof.noveg[[x]], top1=NA, toe1=NA, rast1[[x]], bluff=FALSE, NA))
}
