% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_margins.R
\name{make.margins}
\alias{make.margins}
\title{Extracting the sides / margins of a polygon shapefiles as line shapefiles}
\usage{
make.margins(buff1, shore1, rast1, dir.name)
}
\arguments{
\item{buff1}{\itemize{
\item Either full path including file name and extension .shp of one
polygon buffer shapefile, or a SpatialPolygonsDataFrame object that contains
one polygon shapefile. This is usually the result from bBluff function, one
object from the resulting list.
}}

\item{shore1}{\itemize{
\item Either full path including file name and extension .shp of one
line shapefile, or a SpatialLinesDataFrame object that contains only one line
shapefile. This is usually the result from bCoast function, one object from the
resulting list.
}}

\item{rast1}{\itemize{
\item Either full path including file name and extension (e.g. .tif
for geotiff files) of one DEM raster that was clipped with a buffer or a
RasterLayer object.
}}

\item{dir.name}{\itemize{
\item The full path to the directory where a Margins subdirectory
will be created and the individual buffers sides will be saved as line shapefiles,
one shapefile for each side. If the dir.name given does not exist it will be
created. If dir.name is NA, the shapefiles will not be saved on the disk.
}}
}
\value{
A list of lists with two SpatialLinesDataFrame objects each.
}
\description{
This function transform a polygon buffer shapefile into a line shapefile and
split it in 4 lines, the 2 rounded ends and the 2 usually linear sides. The
2 sides are saved as line shapefiles.
}
\examples{
coast1 <- system.file("extdata", "shore_line.shp", package = "iBluff")
buff1 <- bBluff(coast1, NA)
shore1 <- bCoast(coast1, NA)
dem1 <- system.file("extdata", "dem_demo.tif", package = "iBluff")
rast1 <- clip.rast(dem1, buff1,NA)

n <- length(shore1)
comb.n <-  combn(1:n,1)
marg1 <- apply(comb.n, 2, function(x)
  make.margins(buff1[[x]], shore1[[x]], rast1[[x]], NA))
}
