% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shape.R
\name{make.shape}
\alias{make.shape}
\title{Save a point, line, and smoothed line shapefile}
\usage{
make.shape(tabl, dir.shape, file.name, smooth.win = NA, rast1 = NA)
}
\arguments{
\item{tabl}{\itemize{
\item Is a table with the first 2 columns x and y coordinates or a list
of tables with the first 2 columns x and y coordinates/
}}

\item{dir.shape}{\itemize{
\item Full path of the directory where to save the shapefiles.
}}

\item{file.name}{\itemize{
\item A name used as prefix for all saved shapefiles.
}}

\item{smooth.win}{\itemize{
\item How big the smoothing window is, represented by  the number
of cells to smooth over. If smooth.win=NA (default) no smoothing will take place.
}}

\item{rast1}{\itemize{
\item Either full path including file name and extension (e.g. .tif for geotiff files)
of one DEM raster that was clipped with a buffer or a RasterLayer object. If rast1=NA (default)
then the projection is considered an empty string (""), the raster resolution equals 1 and
the raster name that is added by default to the name of the shapefile is "my_raster".
}}
}
\value{
Does not return a value, saves files to hard drive.
}
\description{
The function transform a table of x and y coordinates in a point and line shapefile
with projection if a raster layer or a full path to a raster is given to read
the desired projection from., otherwise the projection string will be empty.
}
\examples{
# Do not run
# Consider the bluff/cliff top results from get.toptoe, and the list of rasters
# rast1 from clip.rast function. "MyDir" is the full path and the directory where
# the files will be saved.
#
# n6 <- length(top1)
# comb.n6 <- combn(1:n6, 1)
#
# apply(comb.n6, 2, function(x)
#   make.shape(top1[[x]], dir.tops="MyDir", file.name="tops", 10, rast1[[x]]))

}
