% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/malinvaud.R
\name{malinvaud}
\alias{malinvaud}
\title{Malinvaud's test for significance of the CA dimensions}
\usage{
malinvaud(data)
}
\arguments{
\item{data}{Name of the dataset.}
}
\description{
This function allows you to perform the Malinvaud's test, which assesses the significance of the CA dimensions. The function returns both a table in the R console and a plot. The former lists some values, among which the significance of each CA dimension.Dimensions whose p value is below the 0.05 threshold (displayed in RED in the returned plot) are significant.
}
\examples{
data(greenacre_data)
malinvaud(greenacre_data)
}
\keyword{Malinvaud}
\keyword{test}

