% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rows_cntr_scatter.R
\name{rows.cntr.scatter}
\alias{rows.cntr.scatter}
\title{Scatterplot for row categories contribution to dimensions}
\usage{
rows.cntr.scatter(data, x, y)
}
\arguments{
\item{x,y:}{dimensions for which the contributions are reported.}

\item{data:}{name of the dataset.}
}
\description{
This function allows to plot a scatterplot of the contribution of row categories to two selected dimensions. Two references lines (in RED) indicate the threshold above which the contribution can be considered important for the determination of the dimensions. A diagonal line (in BLACK) is just a visual aid to eyeball whether a category is actually contributing more (in relative terms) to either of the two dimensions.
}
\examples{
data(greenacre_data) #load the sample dataset
rows.cntr.scatter(greenacre_data,1,2) #plot the scatterplot for dimensions 1&2
}
\keyword{categories}
\keyword{contribution}
\keyword{row}
\keyword{scatterplot}

