% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rows_corr.R
\name{rows.corr}
\alias{rows.corr}
\title{Rows correlation to the dimensions chart}
\usage{
rows.corr(data, x)
}
\arguments{
\item{data:}{ame of the dataset.}

\item{x:}{dimension for which the row categories correlation is returned.}
}
\description{
This function allows you to calculate the correlation of the row categories to the selected dimensions.
}
\examples{
data(greenacre_data)
rows.corr(greenacre_data, 1) # Returns the correlation of the row categories to the 1 CA dimension.
}
\keyword{correlation}
\keyword{rows}

