% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rows_corr_scatter.R
\name{rows.corr.scatter}
\alias{rows.corr.scatter}
\title{Scatterplot for row categories correlation with dimensions}
\usage{
rows.corr.scatter(data, x, y)
}
\arguments{
\item{x,y:}{dimensions for which the correlations are reported.}

\item{data:}{name of the dataset.}
}
\description{
This function allows to plot a scatterplot of the correlation of row categories with two selected dimensions. A diagonal line (in BLACK) is just a visual aid to eyeball whether a category is actually more correlated (in relative terms) to either of the two dimensions.
}
\examples{
data(greenacre_data) #load the sample dataset
rows.corr.scatter(greenacre_data,1,2) #plot the scatterplot for dimensions 1&2
}
\keyword{categories}
\keyword{correlation}
\keyword{row}
\keyword{scatterplot}

