% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rows_qlt.R
\name{rows.qlt}
\alias{rows.qlt}
\title{Chart of rows quality of the display}
\usage{
rows.qlt(data, x, y)
}
\arguments{
\item{x,y:}{dimensions for which the correlations are returned.}

\item{data:}{name of the dataset.}
}
\description{
This function allows you to calculate the quality of the display of the row categories on pairs of selected dimensions.
}
\examples{
data(greenacre_data)
rows.qlt(greenacre_data,1,2) #Returns the quality of the display of the row categories on the 1&2 dimensions
}
\keyword{quality}
\keyword{rows}

