% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sig_dim_perm.R
\name{sig.dim.perm}
\alias{sig.dim.perm}
\title{Permuted significance of CA dimensions}
\usage{
sig.dim.perm(data, x, y)
}
\arguments{
\item{x,y:}{dimensions whose significance is calculated.}

\item{data:}{name of the dataset.}
}
\description{
This function allows you to calculate the permuted significance of selected CA dimensions. Number of permutation set at 999. Since the function returns a scatterplot, the function requires that two dimensions are entered.
}
\examples{
data(greenacre_data)
sig.dim.perm(greenacre_data, 1,2) #Returns a scatterplot of the permuted inertia of the 1 CA dimension against the permuted inertia of the 2 CA dimension. Observed inertia of the selected dimensions and 95th percentile of the permuted inertias are also displayed for testing the significance of the observed inertias.
}
\keyword{dimensions}
\keyword{permuted}
\keyword{significance}

