% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sig_tot_inertia_perm.R
\name{sig.tot.inertia.perm}
\alias{sig.tot.inertia.perm}
\title{Permuted significance of the CA total inertia}
\usage{
sig.tot.inertia.perm(data)
}
\arguments{
\item{data:}{name of the dataset.}
}
\description{
This function allows you to calculate the permuted significance of CA total inertia. Number of permutation set at 999.
}
\examples{
data(greenacre_data)
sig.tot.inertia.perm(greenacre_data) #Returns the density curve of the permuted total inertia. Observed total inertia and the 95th percentile of the permuted inertia are also displayed for testing the significance of the observed total inertia.
}
\keyword{inertia}
\keyword{permuted}
\keyword{significance}
\keyword{total}

